Remove BuilderContextCreator
